requirements.txt
azure-identity==1.19.0
numpy==2.2.0
pandas==2.2.3
pyodbc==5.2.0
python-decouple==3.8
